Downgrade for use with Android.
Downgrade to Java 7 (1.7), removed Log4j, added StreamSupport backport, some minor fixes.
This helps to run this library on older Java environments, such as Android.

Should be the default when using this library, as using it works perfectly without Java 8 features and log4j.
Keeping it on an older version makes the library way more versatile.

This commit will be followed by some cleaning commits removing the tests and sample projects.
That's because they don't do well without Java 8 and are not necessary for using the library in another project, so they will not be updated for the changes.